requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
pymongo==4.5.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
dnspython==2.4.2
eventlet==0.33.3